@@ -2,23 +2,23 @@ var url = require('url');
 var crypto = require('crypto');
 
 // Check Timestamp Expired or not
-exports.isTimestampExpired = function(timestamp) {
+exports.isTimestampExpired = function (timestamp) {
     return timestamp < parseInt(Date.now() / 1000);
 };
 
 // Encoded Entry
-exports.encodedEntry = function(bucket, key) {
+exports.encodedEntry = function (bucket, key) {
     return exports.urlsafeBase64Encode(bucket + (key ? ':' + key : ''));
 };
 
 // Get accessKey from uptoken
-exports.getAKFromUptoken = function(uploadToken) {
+exports.getAKFromUptoken = function (uploadToken) {
     var sepIndex = uploadToken.indexOf(':');
     return uploadToken.substring(0, sepIndex);
 };
 
 // Get bucket from uptoken
-exports.getBucketFromUptoken = function(uploadToken) {
+exports.getBucketFromUptoken = function (uploadToken) {
     var sepIndex = uploadToken.lastIndexOf(':');
     var encodedPutPolicy = uploadToken.substring(sepIndex + 1);
     var putPolicy = exports.urlSafeBase64Decode(encodedPutPolicy);
@@ -32,28 +32,27 @@ exports.getBucketFromUptoken = function(uploadToken) {
     }
 };
 
-
-exports.base64ToUrlSafe = function(v) {
+exports.base64ToUrlSafe = function (v) {
     return v.replace(/\//g, '_').replace(/\+/g, '-');
 };
 
-exports.urlSafeToBase64 = function(v) {
+exports.urlSafeToBase64 = function (v) {
     return v.replace(/_/g, '/').replace(/-/g, '+');
 };
 
 // UrlSafe Base64 Decode
-exports.urlsafeBase64Encode = function(jsonFlags) {
-    var encoded = new Buffer(jsonFlags).toString('base64');
+exports.urlsafeBase64Encode = function (jsonFlags) {
+    var encoded = Buffer.from(jsonFlags).toString('base64');
     return exports.base64ToUrlSafe(encoded);
 };
 
 // UrlSafe Base64 Decode
-exports.urlSafeBase64Decode = function(fromStr) {
-    return new Buffer(exports.urlSafeToBase64(fromStr), 'base64').toString();
+exports.urlSafeBase64Decode = function (fromStr) {
+    return Buffer.from(exports.urlSafeToBase64(fromStr), 'base64').toString();
 };
 
 // Hmac-sha1 Crypt
-exports.hmacSha1 = function(encodedFlags, secretKey) {
+exports.hmacSha1 = function (encodedFlags, secretKey) {
     /*
    *return value already encoded with base64
    * */
@@ -67,9 +66,9 @@ exports.hmacSha1 = function(encodedFlags, secretKey) {
 // @param requestURI 请求URL
 // @param reqBody    请求Body，仅当请求的 ContentType 为
 //                   application/x-www-form-urlencoded时才需要传入该参数
-exports.generateAccessToken = function(mac, requestURI, reqBody) {
-    var u = url.parse(requestURI);
-    var path = u.path;
+exports.generateAccessToken = function (mac, requestURI, reqBody) {
+    var u = new url.URL(requestURI);
+    var path = u.pathname + u.search;
     var access = path + '\n';
 
     if (reqBody) {
@@ -86,18 +85,18 @@ exports.generateAccessToken = function(mac, requestURI, reqBody) {
 // @param requestURI     请求URL
 // @param reqMethod      请求方法，例如 GET，POST
 // @param reqContentType 请求类型，例如 application/json 或者  application/x-www-form-urlencoded
-// @param reqBody        请求Body，仅当请求的 ContentType 为 application/json 或者 
+// @param reqBody        请求Body，仅当请求的 ContentType 为 application/json 或者
 //                       application/x-www-form-urlencoded 时才需要传入该参数
 exports.generateAccessTokenV2 = function (mac, requestURI, reqMethod, reqContentType, reqBody) {
-    var u = url.parse(requestURI);
-    var path = u.path;
-    var query = u.query;
+    var u = new url.URL(requestURI);
+    var path = u.pathname;
+    var search = u.search;
     var host = u.host;
     var port = u.port;
 
     var access = reqMethod.toUpperCase() + ' ' + path;
-    if (query) {
-        access += '?' + query;
+    if (search) {
+        access += search;
     }
     // add host
     access += '\nHost: ' + host;
@@ -107,7 +106,7 @@ exports.generateAccessTokenV2 = function (mac, requestURI, reqMethod, reqContent
     }
 
     // add content type
-    if (reqContentType && (reqContentType=='application/json' || reqContentType=='application/x-www-form-urlencoded')) {
+    if (reqContentType && (reqContentType == 'application/json' || reqContentType == 'application/x-www-form-urlencoded')) {
         access += '\nContent-Type: ' + reqContentType;
     }
 
@@ -131,7 +130,7 @@ exports.generateAccessTokenV2 = function (mac, requestURI, reqMethod, reqContent
 // @param reqBody      请求Body，仅当请求的ContentType为
 //                     application/x-www-form-urlencoded时才需要传入该参数
 // @param callbackAuth 回调时请求的Authorization头部值
-exports.isQiniuCallback = function(mac, requestURI, reqBody, callbackAuth) {
+exports.isQiniuCallback = function (mac, requestURI, reqBody, callbackAuth) {
     var auth = exports.generateAccessToken(mac, requestURI, reqBody);
     return auth === callbackAuth;
 };
